step by step it is moving forward. the current code has been verified against its Fortran predecessor using 4 CPU configuration and PI mesh. The exchange functions work but MPI4PY does not nicely work with jax arrays and we have to convert them to numpy back and forth. needs to be fixed. At the end we get: Python & JAX: Total ocean surface area is : 339720437497856.0 m^2 Fortran:      Total ocean surface area is : 339720478579588.  m^2
Provides the coordinates are stored using 10 digits this difference shall be ok and we better save time on researching on that.
This code up to now (except for halo exchange) shall be used to read the mesh and store it. Next we will compute derivatives and
check the computational performance JAX vs Fortran.